refactor(services): extraction de l'algorithme de fusion des descriptions dans son propre module
